code-flix/pyproject.toml
[project]
name = "code-flix"
version = "0.1.0"
description = "Python Modal sandbox runner for codeflix workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.3.2",
    "python-dotenv>=1.2.1",
]
